nixos/lib/make-btrfs-fs: replace fakeroot with proot
Version 6.10.1 of `btrfs-progs` reworked the implementation of the `-r`
option in `mkfs.btrfs`. Prior to this version, the file system produced
by `make-btrfs-fs` was owned by root due to the use of `fakeroot`.
However, the changes made to the `-r` option are incompatible with
`fakeroot`.

The aforementioned version of `btrfs-progs` was introduced to nixpkgs in
August 2024, and affects the current implementation of `make-btrfs-fs`.

There is a  bug report in `fakeroot` tracking this issue:
https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=1081976

The resulting file system not being owned by root can cause errors when
attempting to boot, as documented here:
https://github.com/NixOS/nixpkgs/issues/401263

By switching to `proot`, the original behaviour is preserved.